(edit) restructure the home and about components
