--- src/components/About/index.js
@@ -1,3 +1,0 @@
-export { CCDL } from './CCDL'
-export { Overview } from './Overview'
-export { Hero } from './Hero'
