--- src/components/Home/index.js
@@ -1,4 +1,0 @@
-export { Documents } from './Documents'
-export { Examples } from './Examples'
-export { Features } from './Features'
-export { Hero } from './Hero'
