+++ src/pages/about/index.js
@@ -1,7 +1,9 @@
 import { Box } from 'grommet'
 import { useResponsive } from 'hooks/useResponsive'
+import { AboutCCDL } from 'components/AboutCCDL'
+import { AboutHero } from 'components/AboutHero'
+import { AboutOverview } from 'components/AboutOverview'
 import { FixedContainer } from 'components/shared/FixedContainer'
-import { CCDL, Overview, Hero } from 'components/About'
 
 const About = () => {
   const { setResponsive } = useResponsive()
@@ -16,18 +18,18 @@ const About = () => {
       <FixedContainer
         pad={{ horizontal: setResponsive('none', 'medium', 'basex15') }}
       >
-        <Hero />
+        <AboutHero />
       </FixedContainer>
 
       <FixedContainer
         pad={{ horizontal: setResponsive('medium', 'medium', 'basex15') }}
       >
-        <Overview />
+        <AboutOverview />
       </FixedContainer>
       <FixedContainer
         pad={{ horizontal: setResponsive('medium', 'medium', 'basex15') }}
       >
-        <CCDL />
+        <AboutCCDL />
       </FixedContainer>
     </Box>
   )
