+++ src/pages/index.js
@@ -1,7 +1,10 @@
 import { Box } from 'grommet'
 import { useResponsive } from 'hooks/useResponsive'
+import { HomeDocuments } from 'components/HomeDocuments'
+import { HomeExamples } from 'components/HomeExamples'
+import { HomeFeatures } from 'components/HomeFeatures'
+import { HomeHero } from 'components/HomeHero'
 import { SignUpBlock } from 'components/shared/SignUpBlock'
-import { Documents, Examples, Features, Hero } from 'components/Home'
 
 const Home = () => {
   const { setResponsive } = useResponsive()
@@ -14,11 +17,11 @@ const Home = () => {
           bottom: setResponsive('xlarge', 'xlarge', 'basex12')
         }}
       >
-        <Hero />
-        <Features />
+        <HomeHero />
+        <HomeFeatures />
       </Box>
-      <Examples />
-      <Documents />
+      <HomeExamples />
+      <HomeDocuments />
       <SignUpBlock />
     </>
   )
